Move ConfigDocument and ConfigNode into a "parser" package
This should cut down on people wasting time on these when they
are looking for the regular config API; they are for a narrower
use-case.
